refactor: Reorganize Email and Password TextFields
1. Moved Email and Password TextFields to components folder
2. Updated references in Login Form to use the new components.
3. Removed old Email and Password TextField.
