--- src/pages/Login/LoginForm/EmailTextField.tsx
@@ -1,27 +1,0 @@
-import React from 'react'
-import { TextField } from '@mui/material'
-import { UseFormRegister, FieldErrors } from 'react-hook-form'
-
-import { LoginInputType } from './types'
-
-
-// eslint-disable-next-line no-useless-escape
-const EMAIL_REGEX = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-
-export const EmailTextField: React.FC<{ errors: FieldErrors<LoginInputType>, register: UseFormRegister<LoginInputType> }> = ({ errors, register }) => (
-	<TextField
-		id="email-input"
-		label="Email Address"
-		variant="outlined"
-		margin="normal"
-		type="email"
-		{...register('email', {
-			required: 'Email is Required',
-			pattern: {
-				value: EMAIL_REGEX,
-				message: 'Not a valid email address'
-			}
-		})}
-		error={Boolean(errors.email)}
-		helperText={errors.email?.message} />
-)
